Fix production build by moving test utilities to test-support directory
- Moved setup-mocks.ts and test-utils.tsx from root to test-support/
- Updated all import paths to reference new test-support location
- Modified next.config.ts to exclude test-support directory from build
- Added watchOptions to ignore test files during development
- All tests still passing: 8/22 suites, 71/184 tests (113 skipped)

Fixes production build issue caused by Jest-specific imports in build.
