+++ retro-ai/components/board/__tests__/board-presence.test.tsx
@@ -1,5 +1,5 @@
 import { describe, it, expect, beforeEach, jest } from '@jest/globals';
-import { renderWithProviders, screen, mockUseSocketContext } from '../../../test-utils';
+import { renderWithProviders, screen, mockUseSocketContext } from '../../../test-support/test-utils';
 import { BoardPresence } from '../board-presence';
 
 // Mock the presence utils
